From: Don Lewis <Don.Lewis@tsc.tdk.com> Subject: innfeed-users: connection timer patch
If you're running 0.10.1.3 and having problems with assertion failures, the
attached patch should provide quite a bit of relief.  I've been running
something close to this for a week or so, but I just merged this with the
0.10.1.3 snapshot last night.  So far it doesn't look like I broke anything.
I also added fixes for the two problems I hadn't seen that were reported
in the last day (broken pipes cause assertion failures, and bogus 335
responses cause assertion failures).  There's still one more problem
that I've been trying to track down for most of this week, but it only
seems to crop up once every few days.  Anyway, this has been running OK
for the last 9 hours.

This patch is against the version of connection.c in 0.10.1.3 and does not
contain any of the other changes that were made by patches to this version
that have appeared on the list.

NOTE: this patch was applied after the previous patches


git-svn-id: http://inn.eyrie.org/svn/trunk@32 38acf0ee-cff8-0310-b87c-e736704562cb