From: Jeff Garzik <jeff.garzik@spinne.com> Subject: innfeed-users: innfeed: compiling 0.10.1.2 under Linux
Included below is the patch I created after hacking 0.10.1.2 to get
it to compile under Linux.  In the process I cleaned up a few warnings
and small buglets.  Been running for a couple days with no problems.

This compiles and links cleanly on a Linux 2.0[.30] / Slackware 3.x /
libc 5.3.12 system, but should be applicable to nearly all Linux 2.0
and above systems.

Most importantly, it:
        Disables DO_NEED_SYS_SELECT for Linux
        Enables MMAP support for Linux
        Enables NBIO using FCNTL under Linux

        Jeff


git-svn-id: http://inn.eyrie.org/svn/trunk@33 38acf0ee-cff8-0310-b87c-e736704562cb