Added code to syslog NNRPD article and overview timing and data statistics. Now logs amount of article data sent, amount of time taken, number of articles, amount of overview data read, amount of overview data sent, amount of time to send, and number of overview records sent.
Reduced the priority of the "cant getpeername" message to TRACE.  This is
extremely common when servicing dialup clients and is normal.

Added code to catch a SIGPIPE instead of ignoring it.  This will cause NNRP
to log stats and exit if a SIGPIPE is caught.  Previously disconnected
clients would not be detected until a given command completed.

Changed the declaration of Printf to allow the return value of printf to be
used.  This means that we can use the return value to directly measure the
number of bytes sent.

Obtained from:


git-svn-id: http://inn.eyrie.org/svn/trunk@72 38acf0ee-cff8-0310-b87c-e736704562cb