- Changed the dbz api to use a more natural calling convention.  This means that instead of passing datum's in, you pass HASH's and OFFSET_T's in.  This concludes the breaking of dbm compatibility.  Thank god.  This also is the first step towards manipulating everything as HASH's instead of ASCII message-id's.
- Changed his.c to use HASH's as parameters where message-id strings would
have been used before.  This helps eliminate duplicate hashing of the
messageid. The idea is that we'll hash it once and then pass that around
instead of everything parsing the article to pull the message-id out and
then hashing it again.

- Added a HISremember function.  This is a consequence of keeping track of
the article offered as a hash in the CHANNEL structure.  Since we don't have
the ascii message-id around anymore if we reject the article and move into a
CSeatarticle state, then we have to remember the message-id by other means.
What HISremember does is add [hash] in the message-id field, where hash is
the hex representation of the md5.  Only expire, makehistory, and innd had
to be changed.

- Since case-mapping isn't done in dbz anymore (It's in HashMessageID),
renamed fetch and store to dbzfetch and dbzstore.

- Added HashToText and TextToHash to convert hases to ascii and back.
Better than letting everything else choose how to store them.

- Removed all the references to DBZMAXKEY since we don't care about the
length anymore.

- Added function prototypes for everything in his.c  Need to do this for
everything.

Obtained from:


git-svn-id: http://inn.eyrie.org/svn/trunk@105 38acf0ee-cff8-0310-b87c-e736704562cb