This should make nnrpd work correctly with articles that were stored with the storage api.  If OVER_STYLE is set to MMAP then it will mmap the overview indexes when the group command is executed, otherwise it will incore the entire thing.
A new inn.conf variable was added, 'artmmap'.  If this is false then nnrpd
will use QIO to read articles that weren't stored with nnrpd.  Otherwise it
will mmap articles stored in a traditional spool.

I think that's about it for reader support, I think it'll mostly work but it
hasn't been heavily tested.  OVER_STYLE needs to be made runtime
configurable and all the overview code needs to be changed to use the
ARTnumbers array since that's going to be faster than scanning OVERindex
linearly.


git-svn-id: http://inn.eyrie.org/svn/trunk@180 38acf0ee-cff8-0310-b87c-e736704562cb