Add code to CHANreadloop to allow for restricting number of connections per host, vs current -i option of restricting *total* number of connections.
This is preliminary work, that requires you to go into innd/chan.c to
enable and tune the feature.  Will look into making it settable *per host*
next, instead of a global setting.

There is most likely a cleaner/better way of doing this, but at least the
code is there, and is usable...and seems to improve my servers performance
with its current settings by ~3arts/sec incoming


git-svn-id: http://inn.eyrie.org/svn/trunk@238 38acf0ee-cff8-0310-b87c-e736704562cb