From: Russell Vincent <russellv@uk.uu.net>
Here is a patch to let configure find the missing inet_aton
(e.g: on Solaris) and setup the innfeed Makefile with the correct
options based on it. Checked that it does the right thing on
Solaris (missing) and FreeBSD (not missing).

This now allows the new autoconf to create a compilable setup on
Solaris 2.5.1.  Next phase is the testing.  :-)


git-svn-id: http://inn.eyrie.org/svn/trunk@300 38acf0ee-cff8-0310-b87c-e736704562cb