From: Russell Vincent <russellv@uk.uu.net>
This is a mega-patch of the config-data->inn.conf variables.
The process involved creating a new struct variable 'innconf'
and all the inn.conf variables are placed in that. e.g:

 config.data     inn.conf       usage
MAX_ART_SIZE -> maxartsize -> innconf->maxartsize
WANT_TRASH   -> wanttrash  -> innconf->wanttrash

The new options have been added to the inn.conf(5) man page and
the old config.data options removed from files where necessary.
The entries in config.scoreboard were commented out. A full
list of variables available in 'innconf' can be found in
include/libinn.h and lib/getconfig.c.

Not all config.data options have moved across, because not all
are suitable for this change. PATH options are going to be
a lot tricker, as they appear in man pages and scripts.

Compiles and tested for about a week with no problems.


git-svn-id: http://inn.eyrie.org/svn/trunk@354 38acf0ee-cff8-0310-b87c-e736704562cb