From: Russell Vincent <russellv@uk.uu.net>
This patch adds inn.conf and '-I' command-line options to allow
inn to bind to a specific interface IP address. It also honours the
"INND_BIND_ADDRESS" env variable from a similar patch available on
http://www.nntp.sol.net/patches/.

Included in the patch is the addition of 'port' in inn.conf to
specify the INN port to listen on (compliments the existing '-P'
command-line option which can override the inn.conf option).


git-svn-id: http://inn.eyrie.org/svn/trunk@367 38acf0ee-cff8-0310-b87c-e736704562cb