CNFS (cyclic buffer "filesystem") storage manager ugliness.
Added one #include statement near the top of the file to include
"../storage/cnfs/cnfs.h".  I almost certainly shouldn't be doing that, but
I needed the "extern" declarations for the following in main() which is a
bit of out-of-band communication with the CNFS storage manager:

    __CNFS_Write_Allowed = 1;
    __CNFS_Cancel_Allowed = 1;

My sleep-deprived brain can't suggsest a good way around this, so I'll
commit it for now and hope that Alfonse and Guido don't break into my house
this morning to break my kneecaps.


git-svn-id: http://inn.eyrie.org/svn/trunk@431 38acf0ee-cff8-0310-b87c-e736704562cb