From: Katsuhiro Kondou <kondou@nec.co.jp>
Re-thinking of this, I've got another solution.  timehash_next()
should return art even if OpenArticle() fails.  In that case,
art->len is set to 0.  Then makehistory checks to see if art->len
is 0, and if so, makehistory calls SMcancel() if -b is specified.
Attached is the patch that does this.  This patch includes one
more bug fix in OpenArticle() if amount == RETR_BODY.


git-svn-id: http://inn.eyrie.org/svn/trunk@448 38acf0ee-cff8-0310-b87c-e736704562cb