From:             Richard Michael Todd <rmtodd@mailhost.ecn.ou.edu>
1. nnrpd doesn't ever set OVER_MODE to "r" (open the overview files for read),
so OVER_MODE is still set at the default "a".  This means any subsequent
attempts to actually read the overview file will fail, since it's only open
for writing, not reading.

2. In the branch of code in CMDxover that gets hit when StorageAPI is on
but OVERmmap is not, the code doesn't bother to print the article numbers
in front of the overview lines it dumps.


git-svn-id: http://inn.eyrie.org/svn/trunk@472 38acf0ee-cff8-0310-b87c-e736704562cb