nnrpd.READline.patch - 3-13-98 - curt@kcwc.com (Curt Welch)
This patch fixes a small off-by-one problem with the RFC line length
checks of nnrpd.  It makes nnrpd confirm to the 512 octet limit
of RFC 977 for NNTP commands and the son-of-1036 1000 octet limit
for article data in a POST command.  Before this patch, nnrpd's
limit was one character too short for both (511 & 999).

I'm not aware of any problems this caused for the NNTP command
length, but the article line length would in rare cases cause
problems for long References: lines which were trimmed exactly to
the 998 character limit by a newsreader.  The articles would be
rejected on the POST due to the long line.

Patch for nnrp/misc.c applied to the inn 1.7.2 version.  Also
works on inn 1.5.1 and I suspect many other versions of inn.

(Some portions applied by hand by Forrest Cavalier.)


git-svn-id: http://inn.eyrie.org/svn/trunk@509 38acf0ee-cff8-0310-b87c-e736704562cb