From: Richard Michael Todd <rmtodd@mailhost.ecn.ou.edu>
Couple problems here in overchan.  First, at line 246, a copy of the
"Xref" string is allocated, and a pointer stuck in the Xref pointer
variable. Further down, that variable is freed -- but in the intervening code,
the Xref pointer can change, causing free() to get passed a pointer to
the middle of a malloc()ed block, which will give free() indigestion.
Secondly, the DISPOSE(Xref) is in the wrong place anyway; it should be
*inside* the loop, since a new Xref string is allocated each time through
the loop.


git-svn-id: http://inn.eyrie.org/svn/trunk@521 38acf0ee-cff8-0310-b87c-e736704562cb