From: Michael Hall <mhall@riverside.org>
samples/inncheck.in:

Heres a patch for 'inncheck' to get rid of the error messages it
currently gives regarding 'hosts.nntp' (no longer used in INN 2.0)
and a few 'unknown keywords' in 'inn.conf'.
  It just changes all references from 'hosts.nntp' to 'incoming.conf',
and has a dummy subroutine (just returns 1) for checking the syntax
of 'incoming.conf' for now. I also added lines for the 'unknown
keywords' to match the keywords currently in 'inn.conf'.

nnrpdposthost and extendeddbz is still needed (by kondou@nec.co.jp)


git-svn-id: http://inn.eyrie.org/svn/trunk@684 38acf0ee-cff8-0310-b87c-e736704562cb