From: Richard Todd <rmtodd@skywalker.ecn.ou.edu>
When trying to do, e.g., a 'make clean' in the top level directory, the
Makefile
does basically
   for i in $DIRS
   do
      cd $i; make clean ; cd ..
   done
Now, with 'doc/man' and 'doc/sgml' in the directories list, this code does
      cd doc/man; make clean; cd ..
      cd doc/sgml; make clean; cd ..
See the problem? .. from doc/man is "doc", not the top-level dir.  The
following patch changes the top-level Makefile to go into the 'doc' dir
(and not its subdirs) and have the doc/Makefile do the Right Thing for
make clean/install etc.


git-svn-id: http://inn.eyrie.org/svn/trunk@2031 38acf0ee-cff8-0310-b87c-e736704562cb