Import Heath Kehoe's bug fix from CURRENT:
max-connections and hold-time were not getting set to their default
values.  If you did not set them anywhere in incoming.conf, they'd wind
up with uninitialized values.

If you put a hostname: and a hold-time: in the same block, it would bomb
with 'duplicate key'.  This is because the *_CONFIG macros were wrong.
Same would happen with streaming: and noresendid:.


git-svn-id: http://inn.eyrie.org/svn/branches/STABLE-2_3@3133 38acf0ee-cff8-0310-b87c-e736704562cb