Updated to the latest version:  Made completely stand-alone again so that this program can easily be incorporated in other projects that aren't using libinn.  xmalloc, xstrdup, and sysdie included in the source to remove libinn dependencies.  Switched from concat to xmalloc plus strcpy/strcat.
Added test_print_range to print out a range of test numbers as a dash-
separated range rather than a comma-separated list of test numbers, and
tied this into both the per-test reporting and the end failure summary.
Added a count of skipped tests on ok results.  Fixed the reporting of
exit status in the failed test summary to print -- if the test case didn't
exit cleanly and otherwise to print just the exit status rather than the
full status int.


git-svn-id: http://inn.eyrie.org/svn/trunk@3728 38acf0ee-cff8-0310-b87c-e736704562cb