Moved from include/md5.h.  See that file for older revision history.
Remove the extra prototypes and defines corresponding to other parts of
the MD5 library that we don't implement.  Change the names of the MD5
functions and switch to using better types for some of the parameters.
Get rid of various unnecessary typedefs.  Change the names of some of
the context struct members and use struct md5_context instead of MD5_CTX.


git-svn-id: http://inn.eyrie.org/svn/trunk@3891 38acf0ee-cff8-0310-b87c-e736704562cb