Removed the item for breaking out the nnrpd interface code of the authenticators into a library (done!) and fixing innxmit to be able to handle regular files (done!).
Broke out some of the high-priority projects into sections by the major
release of INN that they're expected to be finished by and reorganized
things somewhat.  Added integration of the new controlchan, IPv6, INN Perl
modules, and an audit against the current NNTP draft to the INN 2.4
schedule.  Added stable libraries and installed header files to the INN
2.6 schedule.

Added clean compilation under make warnings and man pages for all libinn
interfaces to high-priority projects.

Added removal of the rest of LEGACY in config.h, fixing paths.h and
nntp.h, and portable/mmap.h to future projects.  Noted that a preliminary
patch for a PAM authenticator was posted to news.software.nntp.

Added asprintf/vasprintf to long-term projects and clarified various other
entries throughout the file.  Added a plea for submissions to the test
suite at the top of the file.


git-svn-id: http://inn.eyrie.org/svn/trunk@4618 38acf0ee-cff8-0310-b87c-e736704562cb