Add three new options to inn.conf:
nfswriter: assume the spool is on NFS mounted storage and use explicit
msync()s to try and force data out in a timely manner; currently only
implemented for cnfs

nfsreader: assume the spool is on NFS mounted storage and use explicit
msync()s to try and invalidate data before using it to ensure up to
dateness; currently only implemented for cnfs

tradindexedmmap: don't use mmap() for tradindexed, read the data
directly.

tradindexedmmap corresponds to what what #ifdef NFSREADER and more
accurately describes what it does.

We also fix a bug in what was the NFSREADER code so we don't allocate
4 times as much ram as we need to read the overview in (the code was
allocating a pointer's worth of storage for every byte it needed to
read)


git-svn-id: http://inn.eyrie.org/svn/trunk@5471 38acf0ee-cff8-0310-b87c-e736704562cb