Move the overview utility declarations into somewhere more public (ov.h), add three routines: overview_index, overview_split and overview_getheader which:
overview_index:
Given an overview header, return the offset of the field within
the overview data.

overview_split:
Given an overview header line, split out a vector pointing at each
of the components (within line), returning a pointer to the
vector.

overview_getheader:

Given an overview vector (from overview_split), return a copy of
the member which the caller is interested in (and must free).


git-svn-id: http://inn.eyrie.org/svn/trunk@5982 38acf0ee-cff8-0310-b87c-e736704562cb