Convert the build method of libstorage and libinnhist to use non-recursive make.
Add a configure probe for whether the compiler can handle -o and -c at the
same time, and use the results to choose either those flags or mv to allow
object files to be built into a directory other than the current one.
Make buildconfig a bit smarter so that all the makefiles of the storage
methods can be eliminated, and add support for inclusion of makefile
fragments to handle special cases for buffindexed and tradindexed.  Fix all
the rules to not recurse.

This change should fix both dependency handling and parallel make for all
of the storage and history methods.

Also add a fallback to -ldb with no special compilation flags for linking
with BerkeleyDB, to handle the current layout in Debian.


git-svn-id: http://inn.eyrie.org/svn/trunk@6031 38acf0ee-cff8-0310-b87c-e736704562cb