Unfortunately, this has now gotten more complicated (thanks, Solaris).
Add two additional modes.  The first, innbind -t <port>, will test using
innbind to bind a socket to that port to see if it works correctly.  The
inability to bind sockets in a setuid subprocess unfortunately can't be
detected at configure time (because we need a setuid helper program), so
this will have to be used at run-time.

The second mode, innbind -e, will both create the sockets and bind them as
specified on the command line, and then exec the program given on the
command line, with the options given there, adding an additional option
-p followed by a comma-separated list of the bound file descriptors.  This
is equivalent to the old inndstart program, but more flexible and doesn't
require ever parsing any configuration files while root.

The only library code used while root is just buffer_*, vector_*, and the
message functions.


git-svn-id: http://inn.eyrie.org/svn/trunk@6835 38acf0ee-cff8-0310-b87c-e736704562cb