The code for changing users was intended to handle the case where rnews was setuid news, but we don't ever install it that way (for reasons that I can't remember at the moment).  Instead, we sometimes install it setuid root.
This should be looked at closer in the future, since I think that setuid
news should be sufficient.  But in any event, setuid(geteuid()) is going
to do nothing useful at all in a setuid root program.  Instead, if either
the real or effective UIDs are 0, figure out the UID of the news user and
setuid to it.  If we're setuid news, don't do anything at all.  I believe
this will work in both configurations.

Based on a problem reported by Matthias Scheler.


git-svn-id: http://inn.eyrie.org/svn/branches/STABLE-2_4@6915 38acf0ee-cff8-0310-b87c-e736704562cb