Implementation of SASL authentication support for nnrpd from Ken Murchison.
As part of this support, the old method for handling TLS configuration
in nnrpd had to be changed, since sasl_config.c was a hacked version of
internal libsasl code and conflicted with library symbols.  The four
configuration options that were in sasl.conf have been moved to inn.conf
(and underscores removed to fit with the rest of the inn.conf options), and
innupgrade has been updated to know how to make that change.


git-svn-id: http://inn.eyrie.org/svn/trunk@6932 38acf0ee-cff8-0310-b87c-e736704562cb