Make sure that buffer_size is always at least as large as QIO_BUFFERSIZE but still a multiple of st_blksize.
Change QIO to do something more sane with lines that are too long.  Rather
than returning too long of a line once and then returning a partial line
on the next try, consume all of the too-long line before returning the
first error.  This way, the next call will still return the next line of
the file.


git-svn-id: http://inn.eyrie.org/svn/branches/STABLE-2_4@6938 38acf0ee-cff8-0310-b87c-e736704562cb